+++ tools/backups2datalad/__init__.py
@@ -1,0 +1,25 @@
+"""
+IMPORTANT NOTE ABOUT GITHUB CREDENTIALS
+
+This script interacts with the GitHub API (in order to create & configure
+repositories), and so it requires a GitHub OAuth token stored in the global Git
+config under `hub.oauthtoken`.  In addition, the script also pushes commits to
+GitHub over SSH, and so an SSH key that has been registered with a GitHub
+account is needed as well.
+
+Maybe TODO:
+    - do not push in here, push will be outside upon success of the entire hierarchy
+
+Later TODOs
+
+- become aware of superdataset, add new subdatasets if created and ran
+  not for a specific subdataset
+- parallelize across datasets or may be better files (would that be possible within
+  dataset?) using DataLad's #5022 ConsumerProducer?
+"""
+
+import logging
+
+DEFAULT_BRANCH = "draft"
+
+log = logging.getLogger("backups2datalad")
